[familytree] multi-choice-option .on('mouseleave') disappear
refactoring/ setting this code file to correctness.
Given the overly hopeful mistake of tucking our main [ajax.accessAPI data scope/function] in a mouseevent for one of the container children, this file needed work.
our mini promiseHell || event-tangling problem was accommodated by pushing results outward from within the scope into various array buckets. These events were based on clickCounters that incremented upon mouseenter and other behaviors.

great problem to explain the "move fast break things quickly" mantra. in our case: tangled chords can cause fires.

The file was cleaned with an update on callback function knowledge, and with a little jquery .bind() to specify our:
[ elements && event relationships]

its valuable to see why: (if hoverCount == 1) {value based behavior} (if count = 2) {other value based behavior}
is a less sophisticated way of unbinding/removeEventListener when appropriate. 

without putting much to test with .bind(), I imagine it would seem more fluid with setTimeout()  ---> especially if a timeout function has a loop or a data fetch that can cause delay/ allowing a .bind() to override.
i was using a vanish(elem.hide()) callback function that was "cutting the async/sync line"  & can imagine .bind() to be a thing to code defensively around.

 [.slice and .includes can be quite the combo: (1) extract with .slice 2) conditional logic with returned slice value) [.setTimeout && .bind()] or .mouseenter&&.bind().   or behavior/.bind() etc.

Leaving original code within [familytreecopy] to examplify "what not to do " 

please pardon the long explanation, I was afraid of doing commits if they were going to reveal a gap in knowledge/skillset.
Got a ways to go but I'd like to be on track with commits and an occassional summary (like this) to fill in a bundle of commits with specifics.