partials/canvas
This was going to be a snake game /  way of routing to the strain. 
When canvas element mine cart ------> intersects with nav bar element $('/mineimg').getBoundingClientRect() change the route. It's just a simple animation in line with the theming.
It was going to be a non user-controlled Left To Right animation.